+++ examples/20_solar_electrolyzer_doe/run_csv_doe.py
@@ -1,7 +1,7 @@
 """Minimal working example from the DOE user guide docs."""
 
-from h2integrate.core.utilities import load_yaml, check_file_format_for_csv_generator
 from h2integrate.core.dict_utils import update_defaults
+from h2integrate.core.file_utils import load_yaml, check_file_format_for_csv_generator
 from h2integrate.core.h2integrate_model import H2IntegrateModel
