+++ examples/26_floris/run_floris_example.py
@@ -1,7 +1,7 @@
 import os
 from pathlib import Path
 
-from h2integrate.core.utilities import load_yaml
+from h2integrate.core.file_utils import load_yaml
 from h2integrate.core.h2integrate_model import H2IntegrateModel
