+++ h2integrate/converters/iron/martin_ore/cost_model.py
@@ -8,7 +8,7 @@ from pathlib import Path
 import numpy as np
 import pandas as pd
 
-from h2integrate.core.utilities import load_yaml
+from h2integrate.core.file_utils import load_yaml
 
 
 CD = Path(__file__).parent
