+++ h2integrate/converters/iron/rosner/cost_model.py
@@ -9,7 +9,7 @@ from pathlib import Path
 import numpy as np
 import pandas as pd
 
-from h2integrate.core.utilities import load_yaml
+from h2integrate.core.file_utils import load_yaml
 from h2integrate.converters.iron.load_top_down_coeffs import load_top_down_coeffs
